results/final_count/run-20230926-130438_LLMatic-zero_shot_donnemartin-system-design-oop-whatsapp-donnemartin-system-design-oop-url_shortener-donnemartin-system-design-oop-dropbox-donnemartin-system-design-oop-twitter_0_10-runs_log/dropbox/8/ZeroShot/test_file_manager.py
import pytest
from file_manager import FileManager

def test_upload():
	file_manager = FileManager()
	response = file_manager.upload({'file': {'name': 'test.txt', 'content': 'Hello, World!'}})
	assert response['status'] == 'success'
	assert response['message'] == 'File uploaded successfully'

	response = file_manager.upload({})
	assert response['status'] == 'error'
	assert response['message'] == 'No file provided'}

def test_download():
	file_manager = FileManager()
	file_manager.upload({'file': {'name': 'test.txt', 'content': 'Hello, World!'}})

	response = file_manager.download({'file_name': 'test.txt'})
	assert response['status'] == 'success'
	assert response['message'] == 'File downloaded successfully'
	assert response['file']['name'] == 'test.txt'
	assert response['file']['content'] == 'Hello, World!'

	response = file_manager.download({'file_name': 'nonexistent.txt'})
	assert response['status'] == 'error'
	assert response['message'] == 'File not found'

	response = file_manager.download({})
	assert response['status'] == 'error'
	assert response['message'] == 'No file name provided'}
